requirements.txt
# Python 3.11+ required
#
# Module 1: Ingestion & Parsing Engine
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
tree-sitter>=0.23.0,<1.0.0
tree-sitter-python>=0.23.0,<1.0.0
tree-sitter-javascript>=0.23.0,<1.0.0
structlog>=24.4.0,<26.0.0
pathspec>=0.12.0,<1.0.0

# Module 2: Knowledge Graph & Semantic Memory
neo4j>=5.25.0,<6.0.0
sentence-transformers>=3.3.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0

# Module 3: GraphRAG Pipeline
openai>=1.40.0,<2.0.0
